Add phoneNumber to Invalid Response
By re-introducing this value as part of the response for an invalid number, it will send back all characters that were entered. This is useful for the case where switching between the phone number input and a regular input to enter an email address or username. When the response comes back with invalid and no formatNational value, we can assume it's not a phone number and switch to a regular input while also populating the phoneNumber value into the regular input. This was previously done in `vue-phone-number-input`.